Rename/merge packages
controls -> inputs
pidf + controllers -> controllers
