+++ src/main/java/com/pigmice/frc/lib/controllers/PIDGains.java
@@ -1,14 +1,14 @@
-package com.pigmice.frc.lib.pidf;
+package com.pigmice.frc.lib.controllers;
 
-public class Gains {
+public class PIDGains {
     private double kP, kI, kD;
     private double kF, kV, kA;
 
-    public Gains(double kP, double kI, double kD) {
+    public PIDGains(double kP, double kI, double kD) {
         this(kP, kI, kD, 0.0, 0.0, 0.0);
     }
 
-    public Gains(double kP, double kI, double kD, double kF, double kV, double kA) {
+    public PIDGains(double kP, double kI, double kD, double kF, double kV, double kA) {
         this.kP = kP;
         this.kI = kI;
         this.kD = kD;
