+++ src/main/java/com/pigmice/frc/lib/inputs/IBooleanSource.java
@@ -1,4 +1,4 @@
-package com.pigmice.frc.lib.controls;
+package com.pigmice.frc.lib.inputs;
 
 public interface IBooleanSource {
     public boolean get();
